--- a/src/Mvc/Mvc.Razor/test/TagHelpers/UrlResolutionTagHelperTest.cs
+++ b/src/Mvc/Mvc.Razor/test/TagHelpers/UrlResolutionTagHelperTest.cs
@@ -173,6 +173,8 @@ public class UrlResolutionTagHelperTest
                    { "/home/index.html ~/second/wontresolve.html" },
                    { "  ~\\home\\index.html" },
                    { "~\\/home/index.html" },
+                   { "  ~" },
+                   { "   " },
                 };
         }
     }