--- a/src/Mvc/Mvc.Razor/test/TagHelpers/UrlResolutionTagHelperTest.cs
+++ b/src/Mvc/Mvc.Razor/test/TagHelpers/UrlResolutionTagHelperTest.cs
@@ -19,6 +19,7 @@ public class UrlResolutionTagHelperTest
             return new TheoryData<string, string>
                 {
                    { "~/home/index.html", "/approot/home/index.html" },
+                   { "~/home/index.html\r\n", "/approot/home/index.html" },
                    { "  ~/home/index.html", "/approot/home/index.html" },
                    { "\u000C~/home/index.html\r\n", "/approot/home/index.html" },
                    { "\t ~/home/index.html\n", "/approot/home/index.html" },