# Use SearchValues throughout dotnet/aspnetcore (part 3) 

TagBuilder using SearchValues.
UrlResolutionTagHelper using SearchValues.

Both are tasks of https://github.com/dotnet/aspnetcore/issues/46484 

## Description TagBuilder 

- Adding further tests and a new benchmark to compare before-after performance
- Updating the implementation of TagBuilder to use SearchValues instead manual operation

## Performance comparison TagBuilder 

Using `TagBuilderBenchmark` and

BenchmarkDotNet=v0.13.0, OS=Windows 10.0.22631
12th Gen Intel Core i7-1255U, 1 CPU, 12 logical and 10 physical cores
.NET SDK=9.0.100-preview.3.24161.2
  [Host]     : .NET 9.0.0 (9.0.24.17301), X64 RyuJIT
  Job-YBJGDV : .NET 9.0.0 (9.0.24.16003), X64 RyuJIT

Server=True  Toolchain=.NET Core 9.0  RunStrategy=Throughput


BEFORE:

|           Method |     Mean |    Error |   StdDev |         Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|---------:|---------:|-------------:|-------:|------:|------:|----------:|
|   ValidFieldName | 23.61 ns | 0.219 ns | 0.194 ns | 42,361,985.3 |      - |     - |     - |         - |
| InvalidFieldName | 63.85 ns | 0.506 ns | 0.473 ns | 15,662,151.8 | 0.0023 |     - |     - |     208 B |

AFTER:

|           Method |      Mean |     Error |    StdDev |          Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |----------:|----------:|----------:|--------------:|-------:|------:|------:|----------:|
|   ValidFieldName |  2.670 ns | 0.0691 ns | 0.0577 ns | 374,546,269.6 |      - |     - |     - |         - |
| InvalidFieldName | 37.499 ns | 0.2485 ns | 0.2203 ns |  26,667,383.3 | 0.0023 |     - |     - |     208 B |

## Description UrlResolutionTagHelper 

- Adding further tests to cover all whitespace characters in question
- Updating the CreateTrimmed method of UrlResolutionTagHelper to use SearchValues. Could not rely on ROS as public APIs of IUriHelper requires a `string`.

## Performance comparison UrlResolutionTagHelper

Although I have compared perf before-after the changes, for the optimistic path when there are no excess the whitespaces, the perf remained the same (gains are negligible). With extra whitespaces there is some perf gain, but the exact measurement would depend on the actual implementation of `IUriHelper` so the results are only "relative" (in my perf tests I used a FakeUriHelper). Hence, I have not committed the benchmark file nor included the results.

Fixes #46484